requirements.txt
Django==1.6.3
PyYAML==3.11
beautifulsoup4==4.3.2
biopython==1.63
ipython==2.0.0
nltk==2.0.4
numpy==1.8.1
wsgiref==0.1.2
